Handle 'read from write port' functionality for CV carts.  Still TODO are the ones that bankswitch in RAM, since I don't know what should happen there.
Added 'exitlauncher' commandline argument, which specifies the functionality
of the 'Exit to launcher' action; should exiting a game actually go back to
the launcher, or exit the program entirely.  Also added a UI item in UIDialog
for this.

Reworked the fullscreen/windowed handling in FrameBuffer, simplifying the
logic for determining whether the SDL surface is actually in fullscreen
or not.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@1752 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/emucore/EventHandler.cxx
@@ -972,8 +972,15 @@ void EventHandler::handleEvent(Event::Type event, int state)
       if((myState == S_EMULATE || myState == S_CMDMENU) && state)
       {
         myOSystem->settings().saveConfig();
-        myOSystem->deleteConsole();
-        myOSystem->createLauncher();
+
+        // Go back to the launcher, or immediately quit
+        if(!myOSystem->settings().getBool("exitlauncher"))
+        {
+          myOSystem->deleteConsole();
+          myOSystem->createLauncher();
+        }
+        else
+          myOSystem->quit();
       }
       return;
